Refactor livechat service to ts
