+++ app/utils/server/lib/getTimezone.ts
@@ -14,14 +14,14 @@ const padOffset = (offset: string | number): string => {
 const guessTimezoneFromOffset = (offset: string | number): string => moment.tz.names()
 	.find((tz) => padOffset(offset) === moment.tz(tz).format('Z').toString()) || moment.tz.guess();
 
-export const getTimezone = (user: { utcOffset: string | number }): string | void | SettingValue => {
+export const getTimezone = (user: { utcOffset?: number }): string | void | SettingValue => {
 	const timezone = settings.get('Default_Timezone_For_Reporting');
 
 	switch (timezone) {
 		case 'custom':
 			return settings.get('Default_Custom_Timezone');
 		case 'user':
-			return guessTimezoneFromOffset(user.utcOffset);
+			return guessTimezoneFromOffset(user?.utcOffset || 0);
 		default:
 			return moment.tz.guess();
 	}
