+++ definition/ILivechatDepartment.ts
@@ -2,14 +2,16 @@ export interface ILivechatDepartment {
 	_id: string;
 	name: string;
 	enabled: boolean;
-	description: string;
+	description?: string;
 	showOnRegistration: boolean;
 	showOnOfflineForm: boolean;
-	requestTagBeforeClosingChat: boolean;
+	requestTagBeforeClosingChat?: boolean;
 	email: string;
-	chatClosingTags: string[];
+	chatClosingTags?: string[];
 	offlineMessageChannelName: string;
 	numAgents: number;
 	_updatedAt?: Date;
 	businessHourId?: string;
+	// optional fields
+	[k: string]: any;
 }
