+++ package.json
@@ -174,6 +174,7 @@
 		"@rocket.chat/ui-kit": "^0.29.0",
 		"@slack/client": "^4.12.0",
 		"@types/lodash": "^4.14.171",
+		"@types/ua-parser-js": "^0.7.36",
 		"adm-zip": "0.4.14",
 		"agenda": "github:RocketChat/agenda#3.1.2",
 		"apn": "2.2.0",
